added vs6 project files for all mods, imported them into the main quake2.dsw.  removed a few unused files from the mods.  fixed a compile error in CTF.  Sezero, check this commit in chaos/c_base.c.  It uses sprintf in WIN32 for LoadMOTD, otherwise it uses strcpy.

+++ ctf/move.c
@@ -527,7 +527,7 @@ SV_NewChaseDir(edict_t *actor, edict_t *enemy, float dist)
 	}
 
 	/* try other directions */
-	if (((rand() & 3) & 1) || (fabsf(deltay) > fabsf(deltax)))
+	if (((rand() & 3) & 1) || (fabs(deltay) > fabs(deltax)))
 	{
 		tdir = d[1];
 		d[1] = d[2];
